+++ pages/producteur/Product.js
@@ -50,7 +50,7 @@ export function ProducteurProductScreen({ route, navigation }) {
             {/* <View>
               <Text style={styles.name}>{product.name}</Text>
               <Text style={styles.price}>
-                {parseInt(product.price).toFixed(2)}€ /{" "}
+                {parseFloat(product.price).toFixed(2)}€ /{" "}
                 {product.pricetype.data.attributes.name
                   .charAt(0)
                   .toUpperCase() +
